% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_cover_letter}
\alias{check_cover_letter}
\title{Check for the two expected RJwrapper files}
\usage{
check_cover_letter(remaining_files)
}
\arguments{
\item{remaining_files}{a vector of the file names in the submission folder}
}
\description{
Check for the two expected RJwrapper files
}
