% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_filenames}
\alias{check_filenames}
\title{Check for the three files that should contain the filename}
\usage{
check_filenames(remaining_files)
}
\arguments{
\item{remaining_files}{a vector of the file names in the submission folder,
after the removal of the RJwrapper tex and pdf files.

Returns the name of the tex file, that should be matched by the .bib and .R}
}
\description{
Check for the three files that should contain the filename
}
