% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_packages_available}
\alias{check_packages_available}
\title{Check that packages mentioned in the text are available on CRAN}
\usage{
check_packages_available(tex)
}
\arguments{
\item{tex}{the tex file read in by readLines}
}
\description{
Check that packages mentioned in the text are available on CRAN
}
