% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_summary}
\alias{check_summary}
\title{Produce a summary of successes, errors and notes}
\usage{
check_summary(path, file = stdout())
}
\arguments{
\item{path}{location of the submission being checked}

\item{file}{the console output directed to the log, using \code{stdout}}
}
\description{
Show a numerical display of the amount of checks that succeeded, resulted in
an error or returned additional information.
}
