% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_wrappers}
\alias{check_wrappers}
\title{Check for the two expected RJwrapper files}
\usage{
check_wrappers(submission_files)
}
\arguments{
\item{submission_files}{a vector of the file names in the submission folder}
}
\description{
Check for the two expected RJwrapper files
}
