% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{initial_checks}
\alias{initial_checks}
\alias{initial_check_article}
\alias{check_title}
\alias{check_section}
\alias{check_abstract_before_intro}
\alias{check_spelling}
\alias{check_wrappers}
\alias{check_filenames}
\alias{check_unnecessary_files}
\alias{check_cover_letter}
\alias{check_proposed_pkg}
\alias{check_packages_available}
\alias{check_summary}
\title{Complete initial checks for submissions.}
\usage{
initial_checks(dic = "en_US")

initial_check_article(path, dic = "en_US")

check_title(tex)

check_section(tex)

check_abstract_before_intro(tex)

check_spelling(tex, dic = "en_US")

check_wrappers(submission_files)

check_filenames(remaining_files)

check_unnecessary_files(submission_files)

check_cover_letter(remaining_files)

check_proposed_pkg()

check_packages_available(tex)

check_summary(path, file = stdout())
}
\arguments{
\item{dic}{the dictionary used for spelling check. See \code{dict} arguent in \code{\link[hunspell:hunspell]{hunspell::hunspell()}}}

\item{path}{location of the submission being checked}

\item{tex}{the tex file read in by readLines}

\item{submission_files}{a vector of the file names in the submission folder}

\item{remaining_files}{a vector of the file names in the submission folder}

\item{file}{the console output directed to the log, using \code{stdout}}
}
\description{
There are several initial checks undertaken for each submission.
These checks reflect the requests made in the author submission guide.
}
\details{
Results will be show in the initial_checks.log file in the submission folder

Set the directory of the articles repository using \code{set_articles_path}.
}
