% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjournal_article.R
\name{create_article}
\alias{create_article}
\title{Create an article template}
\usage{
create_article(dir_path = "paper", file_name = "article")
}
\arguments{
\item{dir_path}{the directory name that hold the .bib, .tex, and .Rmd}

\item{file_name}{the name of teh .Rmd file}
}
\description{
Create an article template
}
