% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{initial_check_article}
\alias{initial_check_article}
\title{A single article check}
\usage{
initial_check_article(path)
}
\arguments{
\item{path}{Location of the article submission folder to check}
}
\description{
There are several initial checks undertaken for a submission.
These checks reflect the requests made in the author submission guide.
}
\details{
Results will be show in the initial_checks.log file in the submission folder

Set the directory of the articles repository using \code{set_articles_path}.
}
