% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\docType{data}
\name{keywords}
\alias{keywords}
\title{Keywords options for R Journal}
\format{
A vector of length 41
}
\usage{
keywords
}
\description{
Available keywords options for R Journal articles
}
\keyword{datasets}
