% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{log_note}
\alias{log_note}
\title{Produce a log file entry for a note}
\usage{
log_note(text, ..., file = stdout(), .envir = parent.frame())
}
\arguments{
\item{text}{Description of the error that occurred}

\item{...}{Additional inputs for text passed to the glue function}

\item{file}{The console output directed to the log, using \code{stdout}}

\item{.envir}{The environment used to find the text string replacements}
}
\description{
Append a line in the log file that details a note
}
\keyword{internal}
