% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-submission.R
\name{prep_submission}
\alias{prep_submission}
\title{Prepare pre-filled fields in the submission form
\code{prep_submission} generate some answers based on the .tex file to fill
the article submission form. You can save the answers if assigned it to an
object.}
\usage{
prep_submission()
}
\value{
a list
}
\description{
Prepare pre-filled fields in the submission form
\code{prep_submission} generate some answers based on the .tex file to fill
the article submission form. You can save the answers if assigned it to an
object.
}
