% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj_pdf_issue.R, R/rj_web_issue.R
\name{rjournal_pdf_issue}
\alias{rjournal_pdf_issue}
\alias{rjournal_web_issue}
\title{R Markdown output formats for R Journal issues}
\usage{
rjournal_pdf_issue(..., render_all = FALSE)

rjournal_web_issue(toc = FALSE, self_contained = FALSE, rnews = FALSE, ...)
}
\arguments{
\item{...}{Arguments passed to \code{distill::distill_article()}.}

\item{render_all}{Re-render all articles in the issue, even if the page numbers have not changed.}

\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
its size).}

\item{rnews}{This issue is from R News.}
}
\value{
an generated R Journal issue
}
\description{
The R Journal is built upon the distill framework with some modifications.
This output format produces the PDF for an R Journal issue.

The R Journal is built upon the distill framework with some modifications.
This output format behaves almost identically to the
\code{distill::distill_article()} format, with some formatting and structural
changes.
}
