% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj_website.R
\name{rjournal_website}
\alias{rjournal_website}
\title{R Markdown site generator for the R Journal}
\usage{
rjournal_website(input, ...)
}
\arguments{
\item{input}{Website directory (or the name of a file within the directory).}

\item{...}{Currently unused.}
}
\description{
R Markdown site generator for Distill websites
}
\keyword{internal}
