% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjtools-package.R
\docType{package}
\name{rjtools-package}
\alias{rjtools}
\alias{rjtools-package}
\title{rjtools: Tools for Preparing, Checking, and Submitting Articles to the R Journal}
\description{
Tools for writing, checking, and submitting articles to the R 
  Journal.
}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Stephanie Kobakian
}

}
\keyword{internal}
