% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-rjournal.R
\name{submit_rjournal}
\alias{submit_rjournal}
\title{Submit a paper to the R Journal}
\usage{
submit_rjournal()
}
\description{
This is a last-minute self-serve  check function, similar to
\code{\link[devtools:release]{devtools::release()}}, for handling R Journal styling that can be difficult
with automatic checks.
}
