% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjournal_article.R
\name{create_article}
\alias{create_article}
\title{Create an Rj article with existing template}
\usage{
create_article(dir_path = "paper", file_name = "article")
}
\arguments{
\item{dir_path}{the directory name that houses the template files}

\item{file_name}{the name of the bib/Rmd file}
}
\description{
Create an Rj article with existing template
}
\details{
see \link\href{\code{vignette("create_article", package = "rjtools")}}{the vignette} for how to create an article
}
