% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjournal_article.R
\name{create_article}
\alias{create_article}
\title{Create an R journal article with the existing template}
\usage{
create_article(file_name = "article", dir_path = here::here())
}
\arguments{
\item{file_name}{the name of the Rmd file}

\item{dir_path}{the directory name that houses the template files}
}
\description{
Create an R journal article with the existing template
}
\details{
See \code{vignette("create_article", package = "rjtools")} for how to create an article
}
