% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjournal_article.R
\name{create_article}
\alias{create_article}
\title{Create an R Journal article with the existing template}
\usage{
create_article(
  name,
  file = xfun::with_ext(name, "Rmd"),
  create_dir = FALSE,
  edit = TRUE
)
}
\arguments{
\item{name}{the name of the Rmd}

\item{file}{File name for the draft}

\item{create_dir}{\code{TRUE} to create a new directory for the document
(the "default" setting leaves this behavior up to the creator of the
template).}

\item{edit}{\code{TRUE} to edit the template immediately}
}
\value{
a created R Journal template in the directory specified by \code{dir_path}
}
\description{
Create an R Journal article with the existing template
}
\details{
See \code{vignette("create_article", package = "rjtools")} for how to create an article
}
