% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjournal_article.R
\name{create_article}
\alias{create_article}
\title{Create an R Journal article with the existing template}
\usage{
create_article(
  name = "test",
  file = xfun::with_ext(name, "Rmd"),
  create_dir = FALSE,
  edit = TRUE
)
}
\arguments{
\item{name}{the name of the Rmd, will default to "test"}

\item{file}{File name for the draft}

\item{create_dir}{\code{TRUE} to create a new directory for the document
(the "default" setting leaves this behavior up to the creator of the
template).}

\item{edit}{\code{TRUE} to edit the template immediately}
}
\description{
Create an R Journal article with the existing template
}
\details{
Outputs an R Journal paper template set of files in the directory specified
by \code{dir_path} or in the project directory, if nothing is specified.
See \code{vignette("create_article", package = "rjtools")} for more details
on how to create and format an article.
}
