% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-submission.R
\name{prepare_submission}
\alias{prepare_submission}
\title{Prepare pre-filled fields in the submission form}
\usage{
prepare_submission(name)
}
\arguments{
\item{name}{the file name you used to create the article. See argument \code{file_name} in \code{create_article()}}
}
\value{
auto-generated answers for some R Journal submission questions
}
\description{
Auto-generate answers to some fields in the R Jorunal submission form
}
