% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjournal_article.R
\name{rjournal_article}
\alias{rjournal_article}
\title{R Markdown output format for R Journal articles}
\usage{
rjournal_article(...)
}
\arguments{
\item{...}{Arguments passed to \code{distill::distill_article()}.}
}
\description{
The R Journal is built upon the distill framework with some modifications.
This output format behaves almost identically to the
\code{distill::distill_article()} format, with some formatting and structural
changes.
}
