% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjournal_article.R
\name{rjournal_web_article}
\alias{rjournal_web_article}
\alias{rjournal_pdf_article}
\title{R Markdown output formats for R Journal articles}
\usage{
rjournal_web_article(...)

rjournal_pdf_article(...)
}
\arguments{
\item{...}{Arguments passed to \code{distill::distill_article()} for web articles,
and \code{rticles::rjournal_article()} for pdf articles.}
}
\description{
The R Journal is built upon the distill framework with some modifications.
This output format behaves almost identically to the
\code{distill::distill_article()} format, with some formatting and structural
changes.
}
