% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj_article.R, R/rj_pdf_article.R,
%   R/rj_web_article.R
\name{rjournal_article}
\alias{rjournal_article}
\alias{rjournal_pdf_article}
\alias{rjournal_web_article}
\title{R Markdown output formats for R Journal articles}
\usage{
rjournal_article(toc = FALSE, self_contained = FALSE, ...)

rjournal_pdf_article(..., self_contained = FALSE, citation_package = "natbib")

rjournal_web_article(toc = FALSE, self_contained = FALSE, ...)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
its size).}

\item{...}{Arguments passed to \code{distill::distill_article()} for web articles,
and \code{rticles::rjournal_article()} for pdf articles.}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{default} if neither package is to be used,
which means citations will be processed via the command
\command{pandoc-citeproc}.}
}
\value{
the rendered R Journal article
}
\description{
The R Journal is built upon the distill framework with some modifications.
This output format behaves almost identically to the
\code{distill::distill_article()} format, with some formatting and structural
changes. The \code{rjournal_article()} format will produce both HTML and PDF
outputs, while the \code{rjournal_pdf_article()} produces only the PDF output.
}
