% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjtools-package.R
\docType{package}
\name{rjtools-package}
\alias{rjtools}
\alias{rjtools-package}
\title{rjtools: Preparing, Checking, and Submitting Articles to the 'R Journal'}
\description{
Create an 'R Journal' 'Rmarkdown' template article, that will generate html and pdf versions of your paper. Check that the paper folder has all the required components needed for submission. Examples of 'R Journal' publications can be found at \url{https://journal.r-project.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rjournal/rjtools}
  \item Report bugs at \url{https://github.com/rjournal/rjtools/issues}
}

}
\author{
\strong{Maintainer}: R Journal Technical Editor \email{r-journal-technical@r-project.org}

Authors:
\itemize{
  \item Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})
  \item Stephanie Kobakian
  \item H. Sherry Zhang \email{huizezhangsh@gmail.com} (\href{https://orcid.org/0000-0002-7122-1463}{ORCID})
  \item Di Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
  \item Simon Urbanek \email{simon.urbanek@R-project.org} (\href{https://orcid.org/0000-0003-2297-1732}{ORCID})
  \item Christophe Dervieux \email{cderv@rstudio.com} (\href{https://orcid.org/0000-0003-4474-2498}{ORCID})
}

}
\keyword{internal}
