% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-rjournal.R
\name{submit_rjournal}
\alias{submit_rjournal}
\title{Submit a paper to the R Journal}
\usage{
submit_rjournal()
}
\description{
This is a final self-check function, similar to
that runs through a checklist on your submission
that cannot be done automatically.
}
