% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-submission.R
\name{zip_paper}
\alias{zip_paper}
\title{Zip your directory for R Journal submission}
\usage{
zip_paper()
}
\value{
a zip file for an R Journal submission
}
\description{
\code{zip_paper} will first check the folder structure with
\code{\link[=check_folder_structure]{check_folder_structure()}} before zipping up everything in the
main directory, excluding the \code{.Rproj} file, if exist.
}
